% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difftime.R, R/telltime.R
\name{add}
\alias{add}
\alias{add}
\title{Add together two numbers.}
\usage{
add(x, y)

add(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}

\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.

The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.

Give me the time.
}
\examples{
add(1, 1)
add(10, 1)
add(1, 1)
add(10, 1)
}

