% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addggregression.R
\name{addggregression}
\alias{addggregression}
\title{Add a regression or loess to an existing ggplot}
\usage{
addggregression(myplot, doregression = T, doloess = F, span = if (doloess
  == T) {     c(0.5, 1, 1.5, 2) }, colour = "black", ...)
}
\arguments{
\item{myplot}{The ggplot object that wants to be annotated with a regression}
}
\value{
Returns the pots with the additions
}
\description{
Add a regression or loess to an existing ggplot
}
\examples{
x=rnomr(1,1,100)
y=rnorm(1,2,100)
x=rnorm(1,3,100)
ggdotscolor(x=x,y=y,colorvar=z) \%>\% addggregression()

}
\seealso{
ggdotscolor
}

