% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.numeric.fator.R
\name{as.numeric.factor}
\alias{as.numeric.factor}
\title{Transform factor to numeric
I do not know why this is not in base package.}
\usage{
\method{as.numeric}{factor}(x)
}
\value{
A numeric vector
}
\description{
Transform factor to numeric
I do not know why this is not in base package.
}
\examples{
as.numeric.factor(as.factor(c(1:5)))
}

