% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_test.R
\name{contrast_test}
\alias{contrast_test}
\title{Quick plot + contrast hypothesis test to compare two populations}
\usage{
contrast_test(numeric1, numeric2, method = "t.test", title = NULL,
  wantpoints = FALSE, wantviolin = TRUE, colorpair = c("darkorchid4",
  "aquamarine4"))
}
\value{
Prints a plot with the given test p-value. If title is provided, it prints into a pdf.
}
\description{
Quick plot + contrast hypothesis test to compare two populations
}
\examples{
group1=rnorm(n = 100,mean = 1,sd = 1)
group1=rnorm(n = 100,mean = 2,sd = 1)
thetimenow<-contrast_test(group1,group2)
}

