% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difftime.R
\name{difftime}
\alias{difftime}
\title{Difference in time 
Given a time point in the past, this function gives you the difference}
\usage{
difftime(timestart)
}
\value{
A time string
}
\description{
Difference in time 
Given a time point in the past, this function gives you the difference
}
\examples{
difference_in_time<-difftime(thetimenow)

}

