% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envir_plot.R
\name{envir_plot}
\alias{envir_plot}
\title{Pretty plot of raster object}
\usage{
envir_plot(raster, objectname = deparse(substitute(raster)), dimcol = 100,
  name = "", vecol = NULL, breaks = format(seq(rangecol[1], rangecol[2],
  len = numbreak), digits = 1), rangecol = c(min(values(raster), na.rm = T),
  max(values(raster), na.rm = T)), numbreak = 5, contrast = 0, pdf = F)
}
\value{
Prints a plot with the given test p-value. If title is provided, it prints into a pdf.
}
\description{
Pretty plot of raster object
}
\examples{
group1=rnorm(n = 100,mean = 1,sd = 1)
group1=rnorm(n = 100,mean = 2,sd = 1)
thetimenow<-contrast_test(group1,group2)
}

