% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc.R
\name{fc}
\alias{fc}
\title{Force character}
\usage{
fc(data.frame)
}
\arguments{
\item{data.frame}{data frame to transform. It must contain only values that can be transformed to character}
}
\value{
A character vector
}
\description{
Force character
}

