% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{fn}
\alias{fn}
\title{Force numeric
For when you have to transform a data.frame to a vector of numbers (e.g. for plotting), to save two long commands}
\usage{
fn(data.frame)
}
\arguments{
\item{data.frame}{data frame to transform. It must contain only values that can be transformed to numeric}
}
\value{
A numeric vector
}
\description{
Force numeric
For when you have to transform a data.frame to a vector of numbers (e.g. for plotting), to save two long commands
}
\examples{

}

