% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.object.name.R
\name{get.object.name}
\alias{get.object.name}
\title{Transform the object name into string
Get the name of the object. This is useful when you want to use the name of the data that is used in a function, to name the output.}
\usage{
get.object.name(x)
}
\value{
A string
}
\description{
Transform the object name into string
Get the name of the object. This is useful when you want to use the name of the data that is used in a function, to name the output.
}
\examples{
thedataname<-getobjectname(mydata)
}

