% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdotscolor.R
\name{ggdotscolor}
\alias{ggdotscolor}
\title{Quick scatter plot with points colored by a 3rd variable}
\usage{
ggdotscolor(data = NULL, x, y, varcol, ylab = NULL, xlab = NULL,
  collab = NULL, mycolors = brewer.pal(10, name = "RdBu"))
}
\value{
A scatter plot from base plot with points colored by the rank of a given variable
}
\description{
Quick scatter plot with points colored by a 3rd variable
}
\examples{
x=rnomr(1,1,100)
y=rnorm(1,2,100)
x=rnorm(1,3,100)
plotwithcolors(x=x,y=y,colorvar=z)

}
\seealso{
ggplot_world_map
}

