% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldmap.R
\name{ggplot_world_map}
\alias{ggplot_world_map}
\title{Create a worldmap}
\usage{
ggplot_world_map(mapcol = "seashell3", backcol = "white",
  countrycol = mapcol, xlim = c(-200, +200), ylim = c(-60, +80))
}
\arguments{
\item{mapcol}{color string of the color of the maps}

\item{backcol}{color string of the background of the plot}

\item{countrycol}{color string of the lines separating countries}

\item{ylim}{}
}
\value{
returns a ggplot object
}
\description{
Create a worldmap
}
\examples{
p<-ggplot_world_map()
p<-geom_point(data=data.frame(lat=41, lon=7), aes(x=lon, y=lat))

}

