% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggregression.R
\name{ggregression}
\alias{ggregression}
\title{Quick plot + contrast hypothesis test to compare two populations}
\usage{
ggregression(x, y, color = "black", xlab = "x", ylab = "y")
}
\arguments{
\item{numeric1}{a numeric vector of group 1}

\item{numeric2}{a numeric vector of group 2}

\item{method}{string, either t.student (parametric) or wilcox.test (nonparametric)}

\item{title}{string with the name of the plot}

\item{wantpoints}{Logical FALSE/TRUE to determine whether points will be overplotted}

\item{wantviolin}{Logical FALSE/TRUE to determine whether you want a violine plot in the background}

\item{The}{colors of the two groups. Default c("darkorchid4","aquamarine4")}
}
\value{
Prints a plot with the given test p-value. If title is provided, it prints into a pdf.
}
\description{
Quick plot + contrast hypothesis test to compare two populations
}
\examples{
group1=rnorm(n = 100,mean = 1,sd = 1)
group1=rnorm(n = 100,mean = 2,sd = 1)
thetimenow<-contrast_test(group1,group2)

}

