% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationship_matrix.R
\name{GRM}
\alias{GRM}
\title{Calculate a genetic relationship matrix}
\usage{
GRM(mat)
}
\arguments{
\item{mat}{A matrix of nrows individuals andncol SNPs. The notation must be allele counts, 0, 1, 2 or NA if missing. No column of individual names.}
}
\description{
Calculate a genetic relationship matrix
}

