% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomvariance.R
\name{h2MCMC}
\alias{h2MCMC}
\title{Proportion of variance of a random factor from a MCMCglmm output object}
\usage{
h2MCMC(mcmcglmmobject, randomname)
}
\arguments{
\item{randomname}{}
}
\description{
Proportion of variance of a random factor from a MCMCglmm output object
}

