% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_eq.R
\name{lm_eq}
\alias{lm_eq}
\title{Quick plot + contrast hypothesis test to compare two populations}
\usage{
lm_eq(y, x)
}
\arguments{
\item{y}{a numeric vector of group 1}

\item{x}{a numeric vector of group 2}

\item{method}{string, either t.student (parametric) or wilcox.test (nonparametric)}

\item{title}{string with the name of the plot}

\item{wantpoints}{Logical FALSE/TRUE to determine whether points will be overplotted}

\item{wantviolin}{Logical FALSE/TRUE to determine whether you want a violine plot in the background}

\item{The}{colors of the two groups. Default c("darkorchid4","aquamarine4")}
}
\value{
Prints a plot with the given test p-value. If title is provided, it prints into a pdf.
}
\description{
Quick plot + contrast hypothesis test to compare two populations
}

