% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.pallete.contrast.R
\name{make.pallete.contrast}
\alias{make.pallete.contrast}
\title{Add contrast to a color palette}
\usage{
make.pallete.contrast(vecol, contrast = 1)
}
\value{
A colorRampPallete object with the desired contast if any
}
\description{
Add contrast to a color palette
}
\examples{
make.pallete.contrast(c("red","orange","yellow","green","blue"), contrast=2)
}

