% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsechrX_XXX.R
\name{makechrX_XXX}
\alias{makechrX_XXX}
\title{Function to transform from chromosome X and position XXX to my standart format ChrX_XXX}
\usage{
makechrX_XXX(chr, pos)
}
\arguments{
\item{chr}{A vector of chromosome numbers.}

\item{pos}{A vector of positions in chromosomes.}

\item{strings}{A vector of strings of ChrX_XXX format name.}
}
\value{
Data frame with the chromosome and snp position corresponding to the names.
Vector of strings with the SNP names in format ChrX_XXX
}
\description{
Function to transform from chromosome X and position XXX to my standart format ChrX_XXX
}
\examples{
parseChrX_XXX(strings=c("Chr4_54135","Chr1_18965"))
makeChrX_XXX(chr=c("4", "1"),pos=c("54135","18965") )
}

