% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeformula.R
\name{makeformula}
\alias{makeformula}
\title{Make formula object from a response and multiple predictor variable names. Useful for glm framework.}
\usage{
makeformula(response, predictors)
}
\arguments{
\item{response}{variable response, only one string.}

\item{predictors}{predicting variables, vector of strings.}
}
\value{
A formula object that can be feed to lm, glm, aov, randomForest, etc.
}
\description{
Make formula object from a response and multiple predictor variable names. Useful for glm framework.
}
\examples{
makeformula(response="phenotype",predictors=c("bio1","bio2"))
}

