% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.se.ci.R
\name{mean.ci}
\alias{mean.ci}
\title{Return a nicely formatted mean(ci) string from a numeric vector}
\usage{
\method{mean}{ci}(x, digits = 3, sep = ",")
}
\arguments{
\item{sep}{}
}
\description{
Return a nicely formatted mean(ci) string from a numeric vector
}

