% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseChrX_XXX.R
\name{parseChrX_XXX}
\alias{parseChrX_XXX}
\title{Function to transform from SNP name format ChrX_XXX to X and XXX; and vice versa}
\usage{
parseChrX_XXX(strings)
}
\arguments{
\item{strings}{A vector of strings of ChrX_XXX format name.}

\item{chr}{A vector of chromosome numbers.}

\item{pos}{A vector of positions in chromosomes.}
}
\value{
Data frame with the chromosome and snp position corresponding to the names.
Vector of strings with the SNP names in format ChrX_XXX
}
\description{
Function to transform from SNP name format ChrX_XXX to X and XXX; and vice versa
}
\examples{
parseChrX_XXX(strings=c("Chr4_54135","Chr1_18965"))
makeChrX_XXX(chr=c("4", "1"),pos=c("54135","18965") )
}

