% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corrplot.R
\name{plot_corrplot}
\alias{plot_corrplot}
\title{Plot correlations of all pairs o variables .}
\usage{
plot_corrplot(dataset, columnpos = NULL, filename = NULL,
  method = "pearson")
}
\arguments{
\item{dataset}{A numeric data frame}

\item{columnpos}{In case you want to compute correlations of a subset of columns. Default NULL}

\item{filename}{Optional parameter, if provided a string, it will output the plot in a pdf in the working directory. Default NULL}

\item{the}{different methods possible in cor function. Default "pearson". Other possibilities are "spearman" or "kendall"}
}
\value{
If a file name is indicated, an output pdf plot. Otherwise, prints to the graphic device.
}
\description{
Plot correlations of all pairs o variables .
}
\examples{
library(MASS)
data(mtcars)
cormatrix_r_p(as.numeric(na.omit(mtcars)))
}

