% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdotscolor.R
\name{plotwithcolors}
\alias{plotwithcolors}
\title{Quick scatter plot with points colored by a 3rd variable}
\usage{
plotwithcolors(x, y, colorvar, palette = c("red", "orange", "yellow", "green",
  "blue"), pch = 19, alpha = 0.5)
}
\arguments{
\item{colorvar}{}

\item{numeric1}{a numeric vector of group 1}

\item{numeric2}{a numeric vector of group 2}
}
\value{
A scatter plot from base plot with points colored by the rank of a given variable
}
\description{
Quick scatter plot with points colored by a 3rd variable
}
\examples{
x=rnorm(n = 100,mean = 1,sd = 1)
y=rnorm(n = 100,mean = 2,sd = 1)
z=rnorm(n = 100,mean = 3,sd = 1)
plotwithcolors(x=x,y=y,colorvar=z)

}
\seealso{
ggdotscolor
}

