% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_corr.R
\name{prune_corr}
\alias{prune_corr}
\title{Iterative prunning of a set of variables by a certain correlation threshold.}
\usage{
prune_corr(thedata, threshold = 0.8, iterations = 10)
}
\arguments{
\item{thedata}{A numeric matrix of variables}

\item{threshold}{The threshold of correlation coefficient. Default 0.8}

\item{iterations}{The algorithm function will stop when no more variables are correlated 
over the specificed threshold or given a number of iterations. This is importante for big
datasets with thousands of variables, which perhaps is not that appropriate for the function.}
}
\description{
Iterative prunning of a set of variables by a certain correlation threshold.
}

