% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{qqplot}
\alias{qqplot}
\title{Produce a qqplot to evaluate p-value inflation.}
\usage{
qqplot(pvals, name = NULL)
}
\arguments{
\item{pvals}{A numeric vectors with p-values}

\item{name}{If outputing plot is desirable, string.}
}
\value{
Either outputs a plot in pdf or prints to the graphic device
}
\description{
Produce a qqplot to evaluate p-value inflation.
}
\examples{
qqplot(pvals=c(0.001,0.05,0.9))
}

