% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize_trays.R
\name{randomize_genotype_trays}
\alias{randomize_genotype_trays}
\title{Quick scatter plot with points colored by a 3rd variable}
\usage{
randomize_genotype_trays(accessions, replicates, rowsize, colsize)
}
\arguments{
\item{accessions}{A vector of IDs of your genotypes}

\item{colsize}{}
}
\value{
A randomized list of genotypes on pots of a tray. You can specify if you also want a pseudo-geographic position of genotypes.
}
\description{
Quick scatter plot with points colored by a 3rd variable
}
\examples{


}

