% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomvariance.R
\name{randomvariance}
\alias{randomvariance}
\title{Get the variance of random effects model}
\usage{
randomvariance(lmod, var = NULL)
}
\arguments{
\item{var}{}
}
\description{
Get the variance of random effects model
}

