% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterplot.R
\name{rasterplot}
\alias{rasterplot}
\title{Pretty plot of raster object}
\usage{
rasterplot(raster, objectname = deparse(substitute(raster)), name = "",
  rangecol = c(min(values(raster), na.rm = T), max(values(raster), na.rm =
  T)), midpoint = NULL, dimcol = 50, numbreak = 10, colbreaks = if
  (is.null(midpoint)) {     pretty(rangecol, dimcol) } else {    
  c(pretty(c(rangecol[1], midpoint), n = round(dimcol/2) - 1), midpoint,
  pretty(c(midpoint, rangecol[2]), n = round(dimcol/2) - 1)) } \%>\% unique(),
  legmid = NULL, legbreaks = if (is.null(legmid)) {     pretty(rangecol, n =
  numbreak - 1) } else {     c(pretty(c(rangecol[1], legmid), n =
  round(numbreak/2)), legmid, pretty(c(legmid, rangecol[2]), n =
  round(numbreak/2))) } \%>\% unique(), vecol = c("cyan", "dodgerblue1",
  "blue", "forestgreen", "orange", "red", "purple"), contrast = 0, pdf = F,
  add = F, plotlegend = T, discrete = F)
}
\arguments{
\item{raster}{A raster layer object}

\item{objectname}{The name of the object plotted or other name to keep track of the pdf}

\item{name}{Title fo the plot}

\item{rangecol}{Range of the color palette}

\item{dimcol}{Number of color breaks in the gradient}

\item{numbreak}{Number of breaks in the color scale}

\item{vecol}{Color vector for the palette}

\item{contrast}{Interger for number of increase contrast iteration}

\item{pdf}{Logical whether to output plot as pedf}

\item{add}{Logical whether this plot will be overplotted on an existing one in the graphic device}

\item{plotlegend}{Logical whether the color legend is plotted. Useful when overlaying two maps.}

\item{breaks}{Breaks used in the color scale}
}
\value{
Prints a plot of a raster object with a nice color palette and without axis (easily changed). If desired, it is outputed as pdf'.
}
\description{
Pretty plot of raster object
}
\examples{



}

