% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sink.plot.R
\name{sink.plot}
\alias{sink.plot}
\title{Print a plot stored in an active binding or object into a pdf.}
\usage{
sink.plot(plot, name = deparse(substitute(plot)))
}
\arguments{
\item{plot}{An active binding or an object of a plot}

\item{name}{Name of the pdf where the plot is printd. Default, activebinding name.}
}
\value{
Either outputs a plot in pdf or prints to the graphic device
}
\description{
Print a plot stored in an active binding or object into a pdf.
}
\examples{
p<-plot(1,1)

}

