% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceNload.R
\name{sourceNload}
\alias{sourceNload}
\title{Load all scripts or objects of a folder}
\usage{
sourceNload(path = ".", type = "script", ending = if (type == "script") { 
     ending = ".R" } else if (type == "object") {     ending = ".RDATA" } else {
      ending = NULL })
}
\arguments{
\item{path}{The string path where the files to be loaded are stored. Default current directory "."}

\item{type}{String that can be either 'script' or 'object'}

\item{ending}{String with the file extension. It is assumed that scripts will end in .R and objects in .RDATA}
}
\value{
Simply loads everything that founds into the global environment
}
\description{
Load all scripts or objects of a folder
}

