% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starpvalue.R
\name{starpvalue}
\alias{starpvalue}
\title{Transform a vector of p-values into ***, **, *, ., ns}
\usage{
starpvalue(x)
}
\arguments{
\item{x}{A numeric vector of p-values}
}
\value{
A character vector of stars that represent simbolically significance
}
\description{
Transform a vector of p-values into ***, **, *, ., ns
}
\examples{
starpvalue(x=c(0.0002354, 0.658))
}

