% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startcluster.R
\name{stopcluster}
\alias{stopcluster}
\title{Stop cluster}
\usage{
stopcluster(cl = cl)
}
\arguments{
\item{cl}{cluster object passed by startcluster}
}
\value{
stops cluster
}
\description{
Stop cluster
}
\examples{
stopcluster(cl)
}

