% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substrRight.R
\name{substrRight}
\alias{substrRight}
\title{Extract characters of a string from the right side of it.}
\usage{
substrRight(x, lastpos, giveright = T)
}
\arguments{
\item{x}{string}

\item{lastpos}{from what position counting in the right you want to extract characters}

\item{giveright}{logical True/False whether you want to extract the strings in the left form that number of rom the right.}
}
\value{
The strings in the left of in the right from a given position counting from the right of a word
}
\description{
Extract characters of a string from the right side of it.
}
\examples{
substrRight(monkey, 2,giveright=T)
substrRight(monkey, 2,giveright=F)
}

