% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telltime.R
\name{telltime}
\alias{telltime}
\title{Tell me the time
This function returns the time}
\usage{
telltime()
}
\value{
A time string
}
\description{
Tell me the time
This function returns the time
}
\examples{
thetimenow<-telltime()

}

