% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooktime.R
\name{tooktime}
\alias{tooktime}
\title{Difference in time 
Given a time point in the past, this function gives you the difference}
\usage{
tooktime(timestart)
}
\value{
A time string
}
\description{
Difference in time 
Given a time point in the past, this function gives you the difference
}
\examples{
difference_in_time<-tooktime(thetimenow)

}

