% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.pdf.R
\name{write.pdf}
\alias{write.pdf}
\title{A customized pdf function}
\usage{
write.pdf(filename, heightmm = 173, widthmm = 173, metricsystem = "mm",
  timestamp = F)
}
\arguments{
\item{metricsystem}{String either "mm" or "inch". Default "mm" as it should be ;P}

\item{file}{filename. Include .pdf ending}

\item{height}{height of the plot}

\item{width}{width of the plot}
}
\value{
outputs a pdf
}
\description{
A customized pdf function
}
\examples{
\dontrun{

write.pdf(filename="myplot",heightmm=173,width=85)
plot(1,1)
dev.off()	
}
}

