% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.tsv.R
\name{write.tsv}
\alias{write.tsv}
\title{Customized write.table function with table separated values.}
\usage{
write.tsv(table, file)
}
\arguments{
\item{table}{data.frame, vector, matrix to write}

\item{file}{filename. Include .tsv ending}
}
\value{
outputs a table
}
\description{
Customized write.table function with table separated values.
}
\examples{
\dontrun{
examplematrix<-matrix(ncol=2, c(1,2,3,4))
write.tsv(examplematrix,"myexample.tsv")
}
}

