% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalparse.R
\name{evalparse}
\alias{evalparse}
\title{Wrap of functions eval and parse for less code writing}
\usage{
evalparse(x)
}
\arguments{
\item{the}{function with the required arguments given as a single string}
}
\value{
Returns whatever the function with its given arguments will return.
}
\description{
Wrap of functions eval and parse for less code writing
}
\examples{
sumof5and4 <- evalparse(paste("sum(",4,"+",5,")") )

}
\seealso{
\code{\link[base]{eval}}, \code{\link[base]{parse}}
}

