% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdotscolor.R
\name{ggdotscolor}
\alias{ggdotscolor}
\title{Quick scatter plot with points colored by a 3rd variable}
\usage{
ggdotscolor(x, y, varcol = NULL, ylab = deparse(substitute(y)),
  xlab = deparse(substitute(x)), collab = deparse(substitute(varcol)),
  mycolors = brewer.pal(10, name = "RdBu"), contrast = 0, ...)
}
\arguments{
\item{x}{a numeric vector of group 1}

\item{y}{a numeric vector of group 2}

\item{varcol}{either a numeric vector (continuous scale plotting) or factor variable (discrete plotting)}

\item{data}{A dataset with 3 columns. It is assumed that 1 is x, 2 is y, 3 is varcol. This substitutes the use of x, y and varcol.}
}
\value{
A scatter plot from base plot with points colored by the rank of a given variable
}
\description{
Quick scatter plot with points colored by a 3rd variable
}
\examples{
x=rnomr(1,1,100)
y=rnorm(1,2,100)
x=rnorm(1,3,100)
ggdotscolor(x=x,y=y,colorvar=z)

}
\seealso{
ggplot_world_map
}

