% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jet.colors.R
\name{jet.colors}
\alias{jet.colors}
\title{Color palete jet colors from Matlab}
\usage{
jet.colors(palette = F)
}
\arguments{
\item{palette}{Logical whether you want to return a palette object or not.}
}
\value{
Either a vector of colors or a palette object

examples
plot(1:10,1:10, col=jet.colors(palette=T))
}
\description{
Color palete jet colors from Matlab
}

