% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{manhattan}
\alias{manhattan}
\title{Quick scatter plot with points colored by a 3rd variable}
\usage{
manhattan(data, chrpalette = c("black", "black", "black", "black", "black"),
  empirical = NULL, type = "cumulative", nameplot = "",
  pdfname = paste0(nameplot, "manhattan", ".pdf"), res = 1e+06)
}
\arguments{
\item{data}{A data.frame or matrix, with colnames 'chr', 'pos', and 'pval'.}

\item{chrpalette}{A vector of colors or single color}

\item{type}{Wether you want each chromosome separatedly plotted or all together. String "cumulative" or "perchromosome". Default cumulative}

\item{nameplot}{data A dataset with 3 columns. It is assumed that 1 is x, 2 is y, 3 is varcol. This substitutes the use of x, y and varcol.}

\item{pdfname}{pdf name to export plot}

\item{res}{Numeric with the resoution to which chromosome positions wants to be round. Default Megabase.}
}
\value{
A scatter plot from base plot with points colored by the rank of a given variable
}
\description{
Quick scatter plot with points colored by a 3rd variable
}

